requirements.txt
streamlit==1.30.0
ultralytics==8.3.198
opencv-python-headless==4.11.0.86
pillow==10.4.0
numpy==1.26.4
torch==2.6.0
torchvision==0.21.0



